requirements.txt
PyPDF2
langgraph
langchain-core
langchain-google-genai
python-dotenv
Pillow